backend/src/modules/places/api.py
from rest_framework import viewsets
from django_filters.rest_framework import DjangoFilterBackend
from .models import Place
from .serializers import PlaceSerializer
from django_app.pagination import CustomPagination
from .filters import PlaceFilter

class PlaceViewSet(viewsets.ModelViewSet):
    queryset = Place.objects.all()
    serializer_class = PlaceSerializer
    pagination_class = CustomPagination
    filter_backends = [DjangoFilterBackend]
    filterset_class = PlaceFilter
