backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Alex Candido", email = "alex.candido.tec@gmail.com"},
]
dependencies = [
    "django>=5.2.4",
    "paho-mqtt>=2.1.0",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.7.0",
    "django-filter>=24.2", # Adicionado para filtragem de API
]

requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
manage = "python src/manage.py"
server = "pdm run manage runserver 0.0.0.0:8000"
start_all_backend = "bash -c \"pdm run server & pdm run manage mqtt_listener &\""
